% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_collardb.R
\name{bootstrap_collarsdb}
\alias{bootstrap_collarsdb}
\title{Bootstrap the collardb database}
\usage{
bootstrap_collarsdb(dirpath = "~/.collardb",
  dbname = "collardb.sqlite3")
}
\arguments{
\item{dirpath}{character}

\item{dbname}{character}
}
\value{
null
}
\description{
Create the collardb database. This function creates the tables,
views, etc. required to efficiently manage GPS telemetry data
}
\examples{
\donttest{
bootstrap_collarsdb()
}
}
