% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collardb_bootstrap.R
\name{build_collardb}
\alias{build_collardb}
\title{Builds the collardb database}
\usage{
build_collardb(dirpath = "~/.collardb", dbname = "collardb.sqlite3")
}
\arguments{
\item{dirpath}{character}

\item{dbname}{character}
}
\value{
null
}
\description{
Create the collardb database. This function creates the tables,
views, etc. required to efficiently manage GPS telemetry data
}
\examples{
\donttest{
build_collardb()
}
}
