% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collardb_bootstrap.R
\name{collardb_bootstrap}
\alias{collardb_bootstrap}
\title{Bootstrap the collardb database}
\usage{
collardb_bootstrap(dirpath = "~/.collardb",
  dbname = "collardb.sqlite3")
}
\arguments{
\item{dirpath}{character}

\item{dbname}{character}
}
\value{
null
}
\description{
Create the collardb database. This function creates the tables,
views, etc. required to efficiently manage GPS telemetry data
}
\examples{
\donttest{
collardb_bootstrap()
}
}
