% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collardb_conn.R
\name{collardb_conn}
\alias{collardb_conn}
\title{create connection to collardb}
\usage{
collardb_conn(db = "~/.collardb/collardb.sqlite3")
}
\arguments{
\item{db}{character}
}
\value{
character
}
\description{
create connection to collardb
}
\examples{
\donttest{
conn <- collardb_conn()
}
}
