% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collardb_delete.R
\name{collardb_delete}
\alias{collardb_delete}
\title{Delete collardb}
\usage{
collardb_delete(dirpath = "~/.collardb", force = FALSE)
}
\arguments{
\item{dirpath}{character}

\item{force}{boolean}
}
\value{
null
}
\description{
Delete collardb
}
\examples{
\donttest{
delete_collardb(force = T)
}
}
