% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_collardb.R
\name{delete_collardb}
\alias{delete_collardb}
\title{Delete collardb}
\usage{
delete_collardb(dirpath = "~/.collardb", force = FALSE)
}
\arguments{
\item{dirpath}{character}

\item{force}{boolean}
}
\value{
null
}
\description{
Delete collardb
}
\examples{
\donttest{
delete_collardb(force = T)
}
}
