% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_animals.R
\name{insert_animals}
\alias{insert_animals}
\title{Insert animal data into collardb database}
\usage{
insert_animals(dat, conn = NULL)
}
\arguments{
\item{dat}{data.frame. This data.frame needs to match the following
vector of column names c('serial_number', 'purchase_date', 'frequency', 'vendor', 'model')}

\item{conn}{a SQLite database connection. Optional if using the default database connection
parameters. Otherwise provide your own database connection}
}
\description{
Insert a data.frame, or an object that inherits from a table, into
the collardb SQLite database.
}
\examples{
\donttest{
insert_animals(animals)
}

}
