% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_device.R
\name{insert_devices}
\alias{insert_devices}
\title{Insert device data into collardb database}
\usage{
insert_devices(dat, conn = NULL)
}
\arguments{
\item{dat}{data.frame. This data.frame needs to match the following
vector of column names c('serial_number', 'purchase_date', 'frequency', 'vendor', 'model')}

\item{conn}{a SQLite database connection. Optional if using the default database connection
parameters. Otherwise provide your own database connection}
}
\description{
Insert a data.frame, or an object that inherits from a table, int
the collardb SQLite database.
}
\examples{
\donttest{
insert_devices(devices)
}


}
